# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.LoginRequest:
    properties:
      email:
        description: メールアドレス
        example: user@example.com
        type: string
      password:
        description: パスワード（6文字以上）
        example: password123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  handler.RegisterRequest:
    properties:
      display_name:
        description: 表示名（省略可能）
        example: Test User
        type: string
      email:
        description: メールアドレス
        example: user@example.com
        type: string
      password:
        description: パスワード（6文字以上）
        example: password123
        minLength: 6
        type: string
      username:
        description: ユーザー名（3-50文字）
        example: testuser
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@noraneko-id.com
    name: Noraneko ID Support
    url: https://noraneko-id.com/support
  description: プライベートサービス開発者向けIDaaS（Identity as a Service）API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://noraneko-id.com/terms
  title: Noraneko ID API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: メールアドレスとパスワードでユーザーログインを行います
      parameters:
      - description: ログイン情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ログイン成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: リクエストエラー
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 認証エラー
          schema:
            additionalProperties: true
            type: object
      summary: ユーザーログイン
      tags:
      - 認証
...
  /auth/profile:
    get:
      description: ログイン中のユーザーのプロフィール情報を取得します
      produces:
      - application/json
      responses:
        "200":
          description: ユーザープロフィール
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 認証エラー
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ユーザーが見つからない
          schema:
            additionalProperties: true
            type: object
      security:
      - SessionAuth: []
      summary: ユーザープロフィール取得
      tags:
      - 認証
  /auth/register:
    post:
      consumes:
      - application/json
      description: 新しいユーザーアカウントを作成します
      parameters:
      - description: ユーザー登録情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ユーザー作成成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: リクエストエラー
          schema:
            additionalProperties: true
            type: object
        "409":
          description: ユーザー既に存在
          schema:
            additionalProperties: true
            type: object
      summary: ユーザー新規登録
      tags:
      - 認証
...
securityDefinitions:
  BearerAuth:
    description: Bearer Token認証。「Bearer 」の後にアクセストークンを指定してください。
    in: header
    name: Authorization
    type: apiKey
  SessionAuth:
    description: セッション認証。session_tokenクッキーを使用します。
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"